+++ GUI-dev/src/app/pages/processstore-details/processstore-details.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { ThemeModule } from '../../@theme/theme.module';
 import { ProcessRatingModule } from '../includes/process-ratings/process-ratings.module';
 import { ProcessStoreDetailsComponent } from './processstore-details.component';
+import {ProcessstoreDetailsModalComponent} from './modal/processstore-details.modal';
 
 @NgModule({
   imports: [
@@ -10,6 +11,10 @@ import { ProcessStoreDetailsComponent } from './processstore-details.component';
   ],
   declarations: [
     ProcessStoreDetailsComponent,
+    ProcessstoreDetailsModalComponent,
+  ],
+  entryComponents: [
+    ProcessstoreDetailsModalComponent,
   ],
 })
 export class ProcessstoreDetailsModule { }