+++ CHANGELOG.md
@@ -4,6 +4,25 @@ This file ONLY contains changes made by the AIM17 dev-team. This means, that thi
 of the forked repository.
 The format is partially based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+##2018-06-26 ([#49](https://github.com/amarbajric/EBUSA-AIM17/pull/120))
+## Added
+- Process-Details
+    - Components are in ../pages/process-details
+    - Displays Process information with example picture
+    - Process can be bought
+        - Via BUY-button
+        - Only available for ORG members
+    - Implementation of buttons to configure or start process after payment
+        - Only available if process is bought
+        - Only available for ORG members
+        - Functionality of methods is not implemented
+    - Modal to approve payment
+- Gateway and Serverconfig
+    - Methods and routes to get / post data from / to backend
+    - Buy / add process to organization
+    - Get processes of organization
+    - Get process by id
+
 ##2018-06-26 ([#115](https://github.com/amarbajric/EBUSA-AIM17/pull/115))
 - Added
   - Endpoint `/api/store/process/<processId>/uploadProcessFile` where a file can be uploaded to a